+++ linter/annotation.go
@@ -18,34 +18,8 @@ func CheckAnnotations(r *resolver.Resolver) []Hint {
 	for _, context := range r.Contexts() {
 		for _, fragment := range context.Fragments {
 			for _, definition := range fragment.Definitions {
-
-				if err := definition.ExpectOnlyOf("external", "Fremdsystem", "error", "Fehler", "Ereignis", "event"); err != nil {
-					res = append(res, &InvalidAnnotation{
-						TypeDef: definition,
-						Error:   err,
-					})
-				}
-
-				if _, err := parser.ParseEventAnnotation(definition); err != nil {
-					res = append(res, &InvalidAnnotation{
-						TypeDef: definition,
-						Error:   err,
-					})
-				}
-
-				if _, err := parser.ParseErrorAnnotation(definition); err != nil {
-					res = append(res, &InvalidAnnotation{
-						TypeDef: definition,
-						Error:   err,
-					})
-				}
-
-				if _, err := parser.ParseExternalSystemAnnotation(definition); err != nil {
-					res = append(res, &InvalidAnnotation{
-						TypeDef: definition,
-						Error:   err,
-					})
-				}
+				_ = definition
+				//TODO
 			}
 		}
 	}